Better fix to the bug here
Fixed the bug with the GetTrailingNumber method, fixing it so that it now properly gets or strips the trailing number of a string even when the non-number part of the string is just one character long.  This is just a slight improvement to the previous fix to it, removing a redundant check on line 1623.